+++ src/com/github/maxstupo/flatengine/gui/GuiButton.java
@@ -42,7 +42,7 @@ public class GuiButton extends AbstractGuiNode {
     }
 
     @Override
-    public boolean update(double delta, boolean shouldHandleInput) {
+    public boolean update(float delta, boolean shouldHandleInput) {
 
         if (shouldHandleInput) {
             isMouseOver = isMouseOver();
